+++ AndroidAsync/src/com/koushikdutta/async/util/ThrottleTimeout.java
@@ -4,7 +4,6 @@ import android.os.Handler;
 
 import com.koushikdutta.async.AsyncServer;
 import com.koushikdutta.async.callback.ValueCallback;
-import com.koushikdutta.async.future.Cancellable;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -12,10 +11,8 @@ import java.util.List;
 /**
  * Created by koush on 7/19/16.
  */
-public class ThrottleTimeout<T> {
-    Handlerish handlerish;
+public class ThrottleTimeout<T> extends TimeoutBase {
     ValueCallback<List<T>> callback;
-    long delay;
     ArrayList<T> values = new ArrayList<>();
     ThrottleMode throttleMode = ThrottleMode.Collect;
 
@@ -32,57 +29,15 @@ public class ThrottleTimeout<T> {
         Meter,
     }
 
-    private interface Handlerish {
-        void post(Runnable r);
-        Object postDelayed(Runnable r, long delay);
-        void removeAllCallbacks(Object cancellable);
-    }
 
     public ThrottleTimeout(final AsyncServer server, long delay, ValueCallback<List<T>> callback) {
-        this.delay = delay;
+        super(server, delay);
         this.callback = callback;
-        this.handlerish = new Handlerish() {
-            @Override
-            public void post(Runnable r) {
-                server.post(r);
-            }
-
-            @Override
-            public Object postDelayed(Runnable r, long delay) {
-                return server.postDelayed(r, delay);
-            }
-
-            @Override
-            public void removeAllCallbacks(Object cancellable) {
-                if (cancellable == null)
-                    return;
-                ((Cancellable)cancellable).cancel();
-            }
-        };
     }
 
     public ThrottleTimeout(final Handler handler, long delay, ValueCallback<List<T>> callback) {
-        this.delay = delay;
+        super(handler, delay);
         this.callback = callback;
-        this.handlerish = new Handlerish() {
-            @Override
-            public void post(Runnable r) {
-                handler.post(r);
-            }
-
-            @Override
-            public Object postDelayed(Runnable r, long delay) {
-                handler.postDelayed(r, delay);
-                return r;
-            }
-
-            @Override
-            public void removeAllCallbacks(Object cancellable) {
-                if (cancellable == null)
-                    return;
-                handler.removeCallbacks((Runnable)cancellable);
-            }
-        };
     }
 
     private void runCallback() {
@@ -119,9 +74,5 @@ public class ThrottleTimeout<T> {
 
     public void setThrottleMode(ThrottleMode throttleMode) {
         this.throttleMode = throttleMode;
-    }
-
-    public void setDelay(long delay) {
-        this.delay = delay;
     }
 }
